Fixed template code according to new OpenAPI specification
