+++ example-requests.http
@@ -1,11 +1,4 @@
-
-###
-GET http://localhost:8080/get
-Content-Type: application/json
-Authorization: Basic user@gmail.com password
-
-###
-GET http://localhost:8080/ads/asddsfsdf/comment
+GET http://localhost:8080/ads/1/comments
 Content-Type: application/json
 Authorization: Basic user@gmail.com password
 
@@ -14,11 +7,8 @@ POST http://localhost:8080/login
 Content-Type: application/json
 
 {
-   "username": "user@gmail.com",
-   "password": "password",
-   "firstName":  " asdad",
-   "lastName": " asdad",
-   "phone":  " asdad"
+  "username": "user@gmail.com",
+  "password": "password"
 }
 
 ###
@@ -26,8 +16,12 @@ POST http://localhost:8080/register
 Content-Type: application/json
 
 {
-  "username": "user12",
-  "password": "password1"
+  "username": "user@gmail.com",
+  "password": "password",
+  "firstName": " asdad",
+  "lastName": " asdad",
+  "phone": " asdad",
+  "role": "USER"
 }
 
 ###
@@ -48,21 +42,13 @@ Authorization: Basic user@gmail.com password
 {
   "phone": "33333",
   "lastName": "Ivanov",
-  "firstName": "Ivan",
-  "email": "user",
-  "id": 1
+  "firstName": "Ivan"
 }
-
-###
-GET http://localhost:8080/users/1
-Content-Type: application/json
-Authorization: Basic user@gmail.com password
 
 ###
 GET http://localhost:8080/ads
 Content-Type: application/json
 Authorization: Basic user@gmail.com password
-
 
 ###
 POST http://localhost:8080/users/set_password
@@ -80,19 +66,8 @@ Content-Type: application/json
 Authorization: Basic user@gmail.com password
 
 {
-  "image": "123",
   "price": 123,
   "description": "123",
-  "pk": 1233,
   "title": "123"
 }
-
-###
-GET http://localhost:8080/ads
-Content-Type: application/json
-Authorization: Basic user@gmail.com password
-
-
-
-
