+++ src/main/java/ru/skypro/homework/dto/Register.java
@@ -3,7 +3,8 @@ package ru.skypro.homework.dto;
 import lombok.Data;
 
 @Data
-public class RegisterReq {
+public class Register {
+
     private String username;
     private String password;
     private String firstName;
