+++ src/main/java/ru/skypro/homework/service/impl/AuthServiceImpl.java
@@ -5,43 +5,43 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.springframework.stereotype.Service;
-import ru.skypro.homework.dto.RegisterReq;
-import ru.skypro.homework.dto.Role;
+import ru.skypro.homework.dto.Register;
 import ru.skypro.homework.service.AuthService;
 
 @Service
 public class AuthServiceImpl implements AuthService {
 
-  private final UserDetailsManager manager;
+    private final UserDetailsManager manager;
+    private final PasswordEncoder encoder;
 
-  private final PasswordEncoder encoder;
-
-  public AuthServiceImpl(UserDetailsManager manager, PasswordEncoder passwordEncoder) {
-    this.manager = manager;
-    this.encoder = passwordEncoder;
-  }
-
-  @Override
-  public boolean login(String userName, String password) {
-    if (!manager.userExists(userName)) {
-      return false;
+    public AuthServiceImpl(UserDetailsManager manager,
+                           PasswordEncoder passwordEncoder) {
+        this.manager = manager;
+        this.encoder = passwordEncoder;
     }
-    UserDetails userDetails = manager.loadUserByUsername(userName);
-    return encoder.matches(password, userDetails.getPassword());
-  }
 
-  @Override
-  public boolean register(RegisterReq registerReq, Role role) {
-    if (manager.userExists(registerReq.getUsername())) {
-      return false;
+    @Override
+    public boolean login(String userName, String password) {
+        if (!manager.userExists(userName)) {
+            return false;
+        }
+        UserDetails userDetails = manager.loadUserByUsername(userName);
+        return encoder.matches(password, userDetails.getPassword());
     }
-    manager.createUser(
-        User.builder()
-            .passwordEncoder(this.encoder::encode)
-            .password(registerReq.getPassword())
-            .username(registerReq.getUsername())
-            .roles(role.name())
-            .build());
-    return true;
-  }
+
+    @Override
+    public boolean register(Register register) {
+        if (manager.userExists(register.getUsername())) {
+            return false;
+        }
+        manager.createUser(
+                User.builder()
+                        .passwordEncoder(this.encoder::encode)
+                        .password(register.getPassword())
+                        .username(register.getUsername())
+                        .roles(register.getRole().name())
+                        .build());
+        return true;
+    }
+
 }
